refactor retriever file locations (#116)
* Move retriever files into a separate directory

* Move source images and fix import paths

* Fix import sorting and missed moving farewell file
