+++ genai-perf/genai_perf/inputs/inputs.py
@@ -19,6 +19,9 @@ from pathlib import Path
 from typing import Any, Dict, List, Tuple
 
 from genai_perf.exceptions import GenAIPerfException
+from genai_perf.inputs.converters.output_format_converter_factory import (
+    OutputFormatConverterFactory,
+)
 from genai_perf.inputs.input_constants import (
     DEFAULT_INPUT_DATA_JSON,
     MINIMUM_LENGTH,
@@ -26,11 +29,8 @@ from genai_perf.inputs.input_constants import (
     OutputFormat,
     PromptSource,
 )
-from genai_perf.inputs.input_retriever_factory import InputRetrieverFactory
 from genai_perf.inputs.inputs_config import InputsConfig
-from genai_perf.inputs.output_format_converter_factory import (
-    OutputFormatConverterFactory,
-)
+from genai_perf.inputs.retrievers.input_retriever_factory import InputRetrieverFactory
 
 
 class Inputs:
