+++ genai-perf/genai_perf/inputs/retrievers/farewell.txt
