+++ genai-perf/genai_perf/inputs/retrievers/source_images/dlss.png
