+++ genai-perf/genai_perf/inputs/retrievers/source_images/h100.jpeg
