+++ genai-perf/tests/test_output_format_converter.py
@@ -16,11 +16,11 @@ import random
 
 import pytest
 from genai_perf.inputs.converters import *
-from genai_perf.inputs.input_constants import ModelSelectionStrategy, OutputFormat
-from genai_perf.inputs.inputs_config import InputsConfig
-from genai_perf.inputs.output_format_converter_factory import (
+from genai_perf.inputs.converters.output_format_converter_factory import (
     OutputFormatConverterFactory,
 )
+from genai_perf.inputs.input_constants import ModelSelectionStrategy, OutputFormat
+from genai_perf.inputs.inputs_config import InputsConfig
 
 
 class TestOutputFormatConverter:
